chore: rename lib/deskflow/key_types => lib/deskflow/KeyTypes
