+++ src/lib/deskflow/IClient.h
@@ -8,8 +8,8 @@
 #pragma once
 
 #include "deskflow/IScreen.h"
+#include "deskflow/KeyTypes.h"
 #include "deskflow/clipboard_types.h"
-#include "deskflow/key_types.h"
 #include "deskflow/mouse_types.h"
 #include "deskflow/option_types.h"
