+++ src/lib/deskflow/Screen.h
@@ -9,8 +9,8 @@
 
 #include "deskflow/DragInformation.h"
 #include "deskflow/IScreen.h"
+#include "deskflow/KeyTypes.h"
 #include "deskflow/clipboard_types.h"
-#include "deskflow/key_types.h"
 #include "deskflow/mouse_types.h"
 #include "deskflow/option_types.h"
