+++ src/lib/server/Config.cpp
@@ -11,8 +11,8 @@
 #include "common/stdistream.h"
 #include "common/stdostream.h"
 #include "deskflow/KeyMap.h"
+#include "deskflow/KeyTypes.h"
 #include "deskflow/XDeskflow.h"
-#include "deskflow/key_types.h"
 #include "deskflow/option_types.h"
 #include "net/XSocket.h"
 #include "server/Server.h"
