+++ src/lib/server/Server.h
@@ -16,9 +16,9 @@
 #include "deskflow/Clipboard.h"
 #include "deskflow/DragInformation.h"
 #include "deskflow/INode.h"
+#include "deskflow/KeyTypes.h"
 #include "deskflow/ServerArgs.h"
 #include "deskflow/clipboard_types.h"
-#include "deskflow/key_types.h"
 #include "deskflow/mouse_types.h"
 #include "deskflow/option_types.h"
 #include "server/Config.h"
